#3 : deleteMember 요청 시 password 입력 설정

+++ bulletin-board/src/main/java/com/solo/bulletin_board/member/service/MemberService.java
@@ -96,9 +96,13 @@ public class MemberService {
         return findMember;
     }
 
-    public void deleteMember(long memberId){
+    public void deleteMember(long memberId, String password){
 
         Member findMember = findVerifiedMember(memberId);
+
+        if(!passwordEncoder.matches(password, findMember.getPassword())){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_PASSWORD_NOT_MATCHED);
+        }
 
         memberRepository.delete(findMember);
     }
